+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Markup "modifyUrl" now goes to cancel page, not contact page
 * Require wp-i18n for using sprintf in javascript
 * "Pick a timeslot" before the timeslots in the select menu
+* Format phone for Dialog Insight
 
 ### 1.4.3: May 24th, 2021
 * Times dropdown must be above header in fixed position